+++ mycodo/mycodo_flask/utils/utils_math.py
@@ -116,6 +116,7 @@ def math_mod(form_mod_math, form_mod_type=None):
         mod_math.name = form_mod_math.name.data
         mod_math.period = form_mod_math.period.data
         mod_math.max_measure_age = form_mod_math.max_measure_age.data
+        mod_math.start_offset = form_mod_math.start_offset.data
 
         measurements = DeviceMeasurements.query.filter(
             DeviceMeasurements.device_id == form_mod_math.math_id.data).all()